+++ src/main/java/com/mrpowergamerbr/loritta/commands/vanilla/administration/AdminUtils.kt
@@ -1,0 +1,62 @@
+package com.mrpowergamerbr.loritta.commands.vanilla.administration
+
+import com.mrpowergamerbr.loritta.commands.CommandContext
+import com.mrpowergamerbr.loritta.utils.Constants
+import com.mrpowergamerbr.loritta.utils.remove
+
+object AdminUtils {
+	suspend fun getOptions(context: CommandContext): AdministrationOptions? {
+		var rawArgs = context.rawArgs
+		rawArgs = rawArgs.remove(0) // remove o usuário
+
+		var reason = rawArgs.joinToString(" ")
+
+		val pipedReason = reason.split("|")
+
+		var usingPipedArgs = false
+		var skipConfirmation = context.config.getUserData(context.userHandle.id).quickPunishment
+		var delDays = 7
+
+		var silent = false
+
+		if (pipedReason.size > 1) {
+			val pipedArgs=  pipedReason.toMutableList()
+			val _reason = pipedArgs[0]
+			pipedArgs.removeAt(0)
+
+			pipedArgs.forEach {
+				val arg = it.trim()
+				if (arg == "force" || arg == "f") {
+					skipConfirmation = true
+					usingPipedArgs = true
+				}
+				if (arg == "s" || arg == "silent") {
+					skipConfirmation = true
+					usingPipedArgs = true
+					silent = true
+				}
+				if (arg.endsWith("days") || arg.endsWith("dias") || arg.endsWith("day") || arg.endsWith("dia")) {
+					delDays = it.split(" ")[0].toIntOrNull() ?: 0
+
+					if (delDays > 7) {
+						context.sendMessage(Constants.ERROR + " **|** " + context.getAsMention(true) + context.locale["SOFTBAN_FAIL_MORE_THAN_SEVEN_DAYS"])
+						return null
+					}
+					if (0 > delDays) {
+						context.sendMessage(Constants.ERROR + " **|** " + context.getAsMention(true) + context.locale["SOFTBAN_FAIL_LESS_THAN_ZERO_DAYS"])
+						return null
+					}
+
+					usingPipedArgs = true
+				}
+			}
+
+			if (usingPipedArgs)
+				reason = _reason
+		}
+
+		return AdministrationOptions(reason, skipConfirmation, silent, delDays)
+	}
+
+	data class AdministrationOptions(val reason: String, val skipConfirmation: Boolean, val silent: Boolean, val delDays: Int)
+}
